# doc/api.yaml
openapi: 3.0.3

info:
  title: WASAText API
  description: |-
    API del progetto WASAText, un clone semplificato di WhatsApp Web.
    Questa prima versione copre l'autenticazione semplificata e la gestione
    del nome utente del profilo autenticato.
  version: 0.1.0
  contact:
    name: WASAText
    email: wasatext@example.com

servers:
  - url: http://localhost:3000
    description: Server di sviluppo locale

tags:
  - name: login
    description: Autenticazione semplificata
  - name: user
    description: Gestione del profilo dell'utente autenticato

security:
  - bearerAuth: []

paths:

  /session:
    post:
      tags: ["login"]
      operationId: doLogin
      summary: Esegue il login dell'utente
      description: |-
        Login semplificato: se l'utente esiste ne restituisce l'identificativo,
        altrimenti lo crea e restituisce l'identificativo appena generato.
      security: []
      requestBody:
        description: Nome utente con cui effettuare l'accesso
        required: true
        content:
          application/json:
            schema:
              type: object
              title: LoginRequest
              description: Credenziali di accesso
              properties:
                name:
                  $ref: "#/components/schemas/Username"
              required: ["name"]
      responses:
        "201":
          description: Utente registrato (o già esistente) e sessione creata
          content:
            application/json:
              schema:
                type: object
                title: LoginResponse
                description: Identificativo dell'utente autenticato
                properties:
                  identifier:
                    $ref: "#/components/schemas/Identifier"
                required: ["identifier"]
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  /users/me/username:
    put:
      tags: ["user"]
      operationId: setMyUserName
      summary: Aggiorna il proprio nome utente
      description: |-
        Sostituisce il nome utente dell'utente autenticato con quello indicato.
        Il nome utente deve restare univoco all'interno del sistema.
      requestBody:
        description: Nuovo nome utente
        required: true
        content:
          application/json:
            schema:
              type: object
              title: ChangeUsernameRequest
              description: Richiesta di cambio nome utente
              properties:
                name:
                  $ref: "#/components/schemas/Username"
              required: ["name"]
      responses:
        "200":
          description: Nome utente aggiornato
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: Nome utente già utilizzato da un altro utente
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500": { $ref: "#/components/responses/InternalServerError" }

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |-
        Autenticazione semplificata: il token da inviare nell'header
        `Authorization: Bearer <identifier>` è l'identificativo utente
        restituito da `doLogin`.

  responses:
    BadRequest:
      description: Richiesta malformata o parametri non validi
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Token di autenticazione mancante o non valido
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InternalServerError:
      description: Errore interno del server
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }

  schemas:

    Identifier:
      title: Identifier
      description: Identificativo numerico univoco di una risorsa
      type: integer
      format: int64
      minimum: 1
      maximum: 9223372036854775807
      example: 42

    Username:
      title: Username
      description: Nome utente univoco, composto da lettere, numeri, punto e underscore
      type: string
      pattern: '^[a-zA-Z0-9_.]{3,16}$'
      minLength: 3
      maxLength: 16
      example: maria_rossi

    PhotoUrl:
      title: PhotoUrl
      description: URL da cui scaricare la foto profilo; assente se non impostata
      type: string
      format: uri
      pattern: '^https?://.*$'
      minLength: 8
      maxLength: 2048
      example: http://localhost:3000/media/photos/42.png

    User:
      title: User
      description: Profilo di un utente
      type: object
      properties:
        identifier: { $ref: "#/components/schemas/Identifier" }
        name: { $ref: "#/components/schemas/Username" }
        photoUrl: { $ref: "#/components/schemas/PhotoUrl" }
      required: ["identifier", "name"]

    Error:
      title: Error
      description: Descrizione di un errore restituito dall'API
      type: object
      properties:
        code:
          type: string
          description: Codice simbolico dell'errore
          pattern: '^[a-z_]{3,32}$'
          minLength: 3
          maxLength: 32
          example: bad_request
        message:
          type: string
          description: Messaggio leggibile che descrive l'errore
          pattern: '^[^\n\r]{1,256}$'
          minLength: 1
          maxLength: 256
          example: nome utente non valido
      required: ["code", "message"]